ADD _ logic for refresh save location / REFACT _ icon swicher
