--- src/components/HeaderNav/WeatherNav/wheaterIconNames.txt
@@ -1,59 +1,0 @@
-Moderate or heavy rain in area with thunder				
-Patchy light rain in area with thunder
-Thundery outbreaks in nearby	
-
-Fog																					
-Mist
-Freezing drizzle						
-Freezing fog								
-
-						
-Cloudy
-Overcast
-
-Clear
-Sunny
-
-Light rain shower	
-Light freezing rain
-Moderate rain								
-Moderate rain at times							
-Light rain
-Light drizzle								
-Patchy light drizzle
-Patchy light rain	
-
-
-Partly Cloudy
-
-Torrential rain shower							
-Moderate or heavy rain shower																							
-Moderate or Heavy freezing rain													
-Heavy rain								
-Heavy rain at times																					
-Heavy freezing drizzle	
-Patchy rain nearby	
-
-Moderate or heavy snow in area with thunder				
-Patchy light snow in area with thunder
-Light snow showers	
-Moderate or heavy sleet showers						
-Light sleet showers
-Moderate or heavy snow showers	
-Heavy snow								
-Patchy heavy snow							
-Moderate snow								
-Patchy moderate snow							
-Light snow								
-Patchy light snow							
-Moderate or heavy sleet							
-Light sleet	
-Blowing snow												
-Patchy freezing drizzle nearby					
-Patchy sleet nearby						
-Patchy snow nearby	
-Blizzard	
-
-377	Moderate or heavy showers of ice pellets	
-374	Light showers of ice pellets
-350	Ice pellets
